64bit32bit/requirements.txt
moviepy
Pillow
IPython